refactor: smaller ui changes
